feat: add SiteSelector plugin
The SiteSelector microservice allows setting access rules per service
identifier by matching a regex against the given attribute. It takes a
configuration object with access rules for each service identifier. Each
access rule consists of an attribute and a match regex. If a rule is
satisfied, the processing of the response continues. If no rule is
satisfied, an error is raised.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>